Unboxing removed

+++ src/androidTest/java/InitialChangeNumberTest.java
@@ -63,7 +63,7 @@ public class InitialChangeNumberTest {
                     long changeNumber = -1;
                     if(mIsFirstChangeNumber) {
                         String path = request.getPath();
-                        changeNumber = Long.valueOf(path.substring(path.indexOf("=") + 1)).longValue();
+                        changeNumber = Long.valueOf(path.substring(path.indexOf("=") + 1));
                         mFirstChangeNumberReceived = changeNumber;
                         mIsFirstChangeNumber = false;
                     }
